kb/ingest/requirements.txt
pypdf
sentence-transformers
psycopg2-binary
numpy
pyyaml
